Refactored ModelBinder classes to make them easier to test
